grouped files into Blueprints
