+++ gifsync/core/templates/privacy.html
