+++ gifsync/extensions.py
@@ -1,6 +1,6 @@
 from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
 
-
 db = SQLAlchemy()
+
 login_manager = LoginManager()
